Refactored the Table Header component.
